fix: extract shared build-info helper, skip git in test env
- Extract memoized getBuildInfo() into src/server/lib/build-info.ts
- Both instrumentation.ts and admin.ts now use the shared helper
- Skip git subprocess fallback when NODE_ENV=test
