+++ src/server/lib/build-info.ts
@@ -1,0 +1,34 @@
+import fs from "fs";
+import path from "path";
+
+interface BuildInfo {
+  version: string;
+  commitSha: string;
+}
+
+let cached: BuildInfo | null = null;
+
+export function getBuildInfo(): BuildInfo {
+  if (cached) return cached;
+
+  let version = "unknown";
+  try {
+    const pkg = JSON.parse(fs.readFileSync(path.resolve(process.cwd(), "package.json"), "utf-8"));
+    version = pkg.version;
+  } catch {}
+
+  let commitSha = "unknown";
+  try {
+    commitSha = fs.readFileSync(path.resolve(process.cwd(), ".commit-sha"), "utf-8").trim();
+  } catch {
+    if (process.env.NODE_ENV !== "test") {
+      try {
+        const { execFileSync } = require("child_process");
+        commitSha = execFileSync("git", ["rev-parse", "--short", "HEAD"], { encoding: "utf-8" }).trim();
+      } catch {}
+    }
+  }
+
+  cached = { version, commitSha };
+  return cached;
+}
